Disabled menu items when in 2d or 3d mode. Moved 2D & 3D button to appBar
